Online mode starts reading logs from the end of file and collect new logs until the process is stopped. Offline mode starts reading logs from the start to the end of file and stops.
The new online-resume mode resumes from where it was stopped.
When a log is "submitted" (recorded to file, displayed to screen or send to syslog),
 a cursor file with the next position is updated.
The cursor file is named <firewall_log_file>.cursor.

Mode should be specified in configuration file (not ONLINE_MODE anymore) :
MODE=OFFLINE
MODE=ONLINE
MODE=ONLINE-RESUME

Or in command-line:
--offline
--online
--online-resume